+++ demo/backend/context_processors.py
@@ -1,5 +1,0 @@
-from django.conf import settings
-
-
-def tutorial(request):
-    return {"HAS_TUTORIAL_APP": "backend.tutorial" in settings.INSTALLED_APPS}