+++ demo/context_processors.py
@@ -1,0 +1,5 @@
+from django.conf import settings
+
+
+def gds_version(request):
+    return {"GDS_VERSION": settings.GDS_VERSION}