+++ src/crispy_forms_gds/__init__.py
@@ -3,8 +3,9 @@ A Django application to add django-crispy-forms template pack and layout
 objects for the GOV.UK Design System.
 
 """
+
 __version__ = "0.3.2"
 __author__ = "Wildfish"
 __email__ = "developers@wildfish.com"
-__license__ = "BSD 3-clause License"
+__license__ = "MIT License"
 __copyright__ = "Copyright (C) 2020-2024 Wildfish Ltd."