+++ docs/conf.py
@@ -35,7 +35,7 @@ author = "Wildfish"
 # The short X.Y version.
 version = "0.3"
 # The full version, including alpha/beta/rc tags.
-release = "0.3.2"
+release = "1.0.0"
 
 # -- General configuration ---------------------------------------------------