+++ CHANGELOG.rst
@@ -11,10 +11,15 @@ and by implication, [Semantic Versioning](http://semver.org/).
 
 Latest
 ------
-* Updated project to use uv.
+* Updated project to use uv
 * Dropped support for Django 3.2 LTS
 * Dropped support for django-crispy-forms 1.x
 * Dropped support for python 3.8, minimum version is now 3.10
+* Dropped the base template, use the `Page Template`_ instead
+* Changed demo site to use `pre-compiled`_ govuk-frontend assets
+
+.. _Page Template: https://design-system.service.gov.uk/styles/page-template/
+.. _pre-compiled: https://frontend.design-system.service.gov.uk/install-using-precompiled-files/
 
 0.3.2 (2024-12-29)
 ------------------